requirements.txt
openai==1.13.3
pygame==2.5.2
python-dotenv==1.0.1